requirements.txt
dj-database-url==0.4.1
Django==1.10.1
django-filter==0.15.0
django-localflavor-us==1.1
djangorestframework==3.4.7
gunicorn==19.6.0
mysql==0.0.1
MySQL-python==1.2.5
wheel==0.24.0
whitenoise==3.2.1
